gpu: the Q8_0 scale plane holds the file's f16 bits (widened in the kernel by cvt; bytes per value 1.125 -> 1.0625, every q8_0 site's device bytes = file bytes, V4.1 q8_0 on the card 7,691,304,960 -> 7,264,010,240), and the heads kernel's short-row path back
- q8f32: scale loads are u16 + half_bits_to_f32 everywhere; q8_0_lane_partial_1col takes rows of at most 128 values (one word per lane) by one test instead of the four-branch ladder
- model/kernels: q8_0_gemv_heads calls q8_0_lane_partial_1col directly; the launcher refuses an x_head_stride off the quad grid
- weights/placement/probe/lookup: DevWeight::Q8_0 and Q8_0Derived carry DeviceTensor<u16>, DevWeight::buffer_bytes() is the measured per-buffer size, CardFormat::Q8_0Planes' scale plane is rows x blocks x 2, probe bytes come from resident_size
- gates: gate_p3 adds q8_0/f32 K = 480 and 992 (m = 1, 8); gate_p6's Q8_OTHER_WORDS 5 -> 6 (q8_0_gemv floor 48, a q8_0_gemv_heads row at 40, PIN with derivation); gate_p10's census counts through the owner; gate_load_v41 reads DevWeight::buffer_bytes(); bench_v41's device bytes through resident_size; placement pins (a) and (b) re-pinned with PIN(2026-09-23) lines

Verdict lines identical to base (forced_exact, sigma, 648 nodes, eager = replay, bench_v41 check 21/21); only byte lines move. Both V4.1 load gates pass with residue 0 B.

Lead A6000 ABAB (base 6d1479f vs this, three rounds each, order rotated): V4.1 token graph (784 nodes) 20.868 -> 19.637 ms (-5.90 %; one-launch wo_a 20.058 -> 18.930), the q8_0 sites 5.4-13.6 % faster on 5.56 % fewer bytes (the derived -0.50 ms assumed only the bytes move); V2-Lite step p50 (depth 6, graph) 4.2904 -> 4.2823 ms (-8.2 us, sd <= 0.07 %), q_nope2 10.02 -> 9.20 us in the p8 profile.

+++ crates/gpu/src/model/lookup.rs
@@ -13,12 +13,12 @@ pub(crate) fn dev_weight<'a>(w: &'a Weights, name: &str) -> Result<&'a DevWeight
 }
 
 /// The resident derived Q8_0 planes named by `name` (`qs` rows x k/4 code
-/// words, `d` rows x k/32 scales): a weight the architecture's plan derived
-/// at load and filed under a `derived.` name.
+/// words, `d` rows x k/32 f16 scales): a weight the architecture's plan
+/// derived at load and filed under a `derived.` name.
 pub(crate) fn q8_derived<'a>(
     w: &'a Weights,
     name: &str,
-) -> Result<(&'a DeviceTensor<u32>, &'a DeviceTensor<f32>), GpuError> {
+) -> Result<(&'a DeviceTensor<u32>, &'a DeviceTensor<u16>), GpuError> {
     match w.get(name) {
         Some(DevWeight::Q8_0Derived { qs, d, .. }) => Ok((qs, d)),
         Some(_) => Err(GpuError::tensor("q8_derived", name, "the derived variant")),
